refactor(server): move the uncaught-capture wiring out of the client class
The capture target, fatal policy and $exception_source stamping now live in
an internal PostHogUncaughtExceptionCapture helper; PostHog.setup only calls
install(). A denied handler installation logs and leaves the client running
with capture disabled instead of failing setup. Changesets trimmed to
customer-facing one-liners.
