+++ .changeset/server-uncaught-exceptions.md
@@ -2,6 +2,4 @@
 'posthog-server': minor
 ---
 
-Opt in to capturing uncaught JVM exceptions on the server SDK via `PostHogConfig.captureUncaughtExceptions` (also on the config `Builder`). When enabled, `PostHog` installs a global `Thread.defaultUncaughtExceptionHandler` on setup that captures the crashing exception as an unhandled `$exception` event (mechanism `onuncaughtexception`, `$exception_source: jvm.uncaught_exception_handler`; `$exception_level` is `fatal` for a main-thread crash and `error` for a worker-thread one, whose death the process survives), flushes, and then delegates to the previously registered handler; the handler is removed again on `close()`. Unlike the Android SDK this is gated purely on the local flag — the server SDK never fetches remote config. A fatal `$exception` event takes a dedicated queue path (keyed off the same fatal-event marker the Android SDK's queue uses): the enqueue and the send run as one ordered task on the queue executor, draining the queue batch by batch and ignoring `flushAt`, so a backlog larger than one batch cannot strand the crash event, and the crashing thread blocks on it for at most a bounded timeout (2s). Delivery stays best-effort, the same guarantee class as the Android SDK: the drain can hit the timeout and each HTTP attempt can fail or be cut short by an immediate hard exit.
-
-The queue's periodic flush timer is now a daemon thread (matching the core SDK's queue), so a set-up client no longer keeps a finished — or crashed — JVM alive until `close()` is called.
+Add `captureUncaughtExceptions`: opt in to capturing uncaught JVM exceptions as error tracking events, with a best-effort flush before the process exits. The SDK's flush timer is now a daemon thread and no longer keeps a finished JVM alive until `close()`.
